# app/mgmt/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "此文档由gin-swagger自动生成",
        "title": "xxim-server HTTP API 文档",
        "contact": {
            "name": "showurl"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:6799",
    "basePath": "/api",
    "paths": {
...
        "/ms/add/role": {
            "post": {
                "description": "使用此接口新增角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理系统角色相关接口"
                ],
                "summary": "新增角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mshandler.updateMSRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.AddMSRoleResp"
                        }
                    }
                }
            }
        },
...
        "/ms/delete/role": {
            "post": {
                "description": "使用此接口删除角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理系统角色相关接口"
                ],
                "summary": "删除角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.DeleteMSRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.DeleteMSRoleResp"
                        }
                    }
                }
            }
        },
...
        "/ms/get/role/detail": {
            "post": {
                "description": "使用此接口获取角色详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理系统角色相关接口"
                ],
                "summary": "获取角色详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.GetMSRoleDetailReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/mshandler.getAllMSRoleListRespRole"
                        }
                    }
                }
            }
        },
        "/ms/get/role/list/all": {
            "post": {
                "description": "使用此接口获取所有角色列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理系统角色相关接口"
                ],
                "summary": "获取所有角色列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.GetAllMSRoleListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/mshandler.getAllMSRoleListResp"
                        }
                    }
                }
            }
        },
...
        "/ms/update/role": {
            "post": {
                "description": "使用此接口更新角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理系统角色相关接口"
                ],
                "summary": "更新角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mshandler.updateMSRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.UpdateMSRoleResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mshandler.getAllMSRoleListResp": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "pageNo": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mshandler.getAllMSRoleListRespRole"
                    }
                }
            }
        },
        "mshandler.getAllMSRoleListRespRole": {
            "type": "object",
            "properties": {
                "apiPaths": {
                    "description": "关联接口",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "string"
                },
                "isDisable": {
                    "description": "是否禁用: [0=否, 1=是]",
                    "type": "boolean"
                },
                "member": {
                    "description": "成员数量",
                    "type": "integer"
                },
                "menus": {
                    "description": "关联菜单",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "remark": {
                    "description": "角色备注",
                    "type": "string"
                },
                "sort": {
                    "description": "角色排序",
                    "type": "integer"
                },
                "updateTime": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "mshandler.updateMSRoleReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "description": "公共请求参数",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pb.CommonReq"
                        }
                    ]
                },
                "role": {
                    "description": "角色信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mshandler.updateMSRoleReqRole"
                        }
                    ]
                }
            }
        },
        "mshandler.updateMSRoleReqRole": {
            "type": "object",
            "properties": {
                "apiPathIds": {
                    "description": "接口ID列表",
                    "type": "string"
                },
                "apiPaths": {
                    "description": "接口列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "description": "角色ID",
                    "type": "string"
                },
                "isDisable": {
                    "description": "是否禁用",
                    "type": "boolean"
                },
                "menuIds": {
                    "description": "菜单ID列表",
                    "type": "string"
                },
                "menus": {
                    "description": "菜单列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "remark": {
                    "description": "角色备注",
                    "type": "string"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                }
            }
        },
...
        "pb.AddMSRoleResp": {
            "type": "object",
            "properties": {
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
...
        "pb.CommonReq": {
            "type": "object",
            "properties": {
                "appVersion": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "deviceId": {
                    "type": "string"
                },
                "deviceModel": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "osVersion": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "userAgent": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "pb.CommonResp": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/pb.CommonResp_Code"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "pb.CommonResp_Code": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                7,
                8,
                9,
                10
            ],
            "x-enum-comments": {
                "CommonResp_AlertError": "alert 错误 只有一个确定按钮",
                "CommonResp_AuthError": "鉴权错误 // 应该退出登录",
                "CommonResp_ForbiddenError": "无权限",
                "CommonResp_InternalError": "内部错误",
                "CommonResp_NeedSecondPasswordError": "需要输入二级密码/设置二级密码",
                "CommonResp_RequestError": "请求错误",
                "CommonResp_RetryError": "alert 错误 有一个取消按钮 和一个重试按钮",
                "CommonResp_ToastError": "toast 错误 只有 message",
                "CommonResp_UnknownError": "未知 error"
            },
            "x-enum-varnames": [
                "CommonResp_Success",
                "CommonResp_UnknownError",
                "CommonResp_InternalError",
                "CommonResp_RequestError",
                "CommonResp_AuthError",
                "CommonResp_ToastError",
                "CommonResp_AlertError",
                "CommonResp_RetryError",
                "CommonResp_ForbiddenError",
                "CommonResp_NeedSecondPasswordError"
            ]
        },
...
        "pb.DeleteMSRoleReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "pb.DeleteMSRoleResp": {
            "type": "object",
            "properties": {
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
...
        "pb.GetAllMSRoleListReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "page": {
                    "$ref": "#/definitions/pb.Page"
                }
            }
        },
...
        "pb.GetMSRoleDetailReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "pb.Page": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "pb.UpdateMSRoleResp": {
            "type": "object",
            "properties": {
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
...
    }
}